Consolidate kotlin.Result JNI helpers into KotlinResult
Previously `EffectsBridges.cs` cached `kotlin/ResultKt` (to construct
`Result.Failure` via `createFailure`) while `SuspendBridge.cs`
separately cached `kotlin/Result$Failure` (to read the `exception`
field on a failed resume). Both files duplicated the
dotnet/java-interop#1440 explanation and lived in places that had
nothing else to do with `kotlin.Result`.

Fold both halves into a single internal `KotlinResult` static class
with three operations:

- `CreateFailure(Throwable)` — boxes a Throwable into Result.Failure
  for `Continuation.resumeWith` (used by `LaunchedEffectBody`).
- `IsFailure(IntPtr)` — predicate over a resumed Result handle
  (used by `SuspendBridge`).
- `ExtractException(Java.Lang.Object)` — reads `Result.Failure.exception`
  (used by `SuspendBridge`).

One copy of the `dotnet/java-interop#1440` comment, one
`FindClass("kotlin/ResultKt")` call, one `FindClass("kotlin/Result$Failure")`
call, and the create-vs-read symmetry is now visible from a single
file. When #1440 lands the entire file collapses to direct calls into
the bound `ResultKt`.
